Move ai folder to lib/ (#542)
