+++ lib/ai/custom-middleware.ts
