+++ lib/ai/index.ts
